mc rate control: add precession compensation feedforward
Single (rigid) rotor aircrafts have their dynamics dominated by the
precession torque. Compensating for this effect by a feedforward greatly
improves stability.
Multirotors can also benefit from this feature when the difference in
RPM between the CW and CCW rotors is large (the sum of the angular
momentum of the rotors is non zero).