VTOL: add option for seperate height error threshold right after front transition
Failure to the fixed-wing acutators of a VTOLs most likely show immediately after
a transition to fixed-wing is complete. Thus it makes sense to set the threshold
for the integrated height rate error lower than in the fixed-wing flight portion later
on.

VT_QC_HR_E_I_TRN is used within the first 10s after a transition to FW is over and
it is set to a positive value, otherwise VT_QC_HR_ERROR_I is also used in this part.

Signed-off-by: Silvan Fuhrer <silvan@auterion.com>